requirements.txt
requests
pytest
selenium
allure-pytest
Appium-Python-Client
pytest-xdist
pip install --editable
pip install pytest-parallel
shutil